# Conflicts:
#	pom.xml
#	src/main/java/com/calendar/project/controller/EventController.java
#	src/main/java/com/calendar/project/controller/JSONController.java
#	src/main/java/com/calendar/project/controller/UserController.java
#	src/main/java/com/calendar/project/dao/EventDao.java
#	src/main/java/com/calendar/project/dao/impl/UserDaoImpl.java
#	src/main/java/com/calendar/project/model/Event.java
#	src/main/java/com/calendar/project/service/EventService.java
#	src/main/java/com/calendar/project/service/impl/EventServiceImpl.java
#	src/main/java/com/calendar/project/service/impl/UserServiceImpl.java
#	src/main/resources/database_postgres.sql
#	src/main/webapp/static/createEvent.jsp
#	src/main/webapp/static/events.jsp
#	src/main/webapp/static/index.jsp
#	src/main/webapp/static/userControlPanel.jsp
#	src/main/webapp/static/welcome.jsp

+++ src/main/java/com/calendar/project/dao/EventDao.java
@@ -11,7 +11,7 @@ public interface EventDao {
 
     List<Event> getAllEvents();
 
-    Event getEvent(long eventId);
+    Event getEvent(int eventId);
 
     List<Event> getEventsByUser(long userId);
